Make timer on Trigger signed on TR4+
If the level version is TR4 or 5 then the `timer` field should be a signed value.
Make this change and add a test.
Closes #926

+++ trview.app/Mocks/Elements/ITrigger.h
@@ -21,7 +21,7 @@ namespace trview
             MOCK_METHOD(TriggerType, type, (), (const, override));
             MOCK_METHOD(bool, only_once, (), (const, override));
             MOCK_METHOD(uint16_t, flags, (), (const, override));
-            MOCK_METHOD(uint8_t, timer, (), (const, override));
+            MOCK_METHOD(int16_t, timer, (), (const, override));
             MOCK_METHOD(uint16_t, sector_id, (), (const, override));
             MOCK_METHOD(const std::vector<Command>, commands, (), (const, override));
             MOCK_METHOD(void, set_triangles, (const std::vector<TransparentTriangle>&), (override));
